# Use cached trials in `TPESampler`'s `sample_relative`